added transmit module:
this is a module that takes in 8 bits audio data at rate 12khz. It then
follows a state machine in order to send encoded information similar to that
of the ir protocol in lab03. The protocol involves a sync frame followed by
specific timing of lows and high for encoding of "0" and "1". It should
be able to send out 8 bits at 12 khz rate.